requirements.txt
alembic==1.18.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
cffi==2.0.0
click==8.3.1
cryptography==46.0.4
ecdsa==0.19.1
fastapi==0.128.0
greenlet==3.3.1
h11==0.16.0
httpx==0.28.1
httptools==0.7.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
psycopg2-binary==2.9.9
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
bcrypt==4.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
slowapi==0.1.9
cachetools==7.0.5
amadeus>=9.0.0
requests==2.32.3
sentry-sdk==2.58.0
